@Zoxc 2023-03-03
No. It kind of looks like a problem with `rustc-perf`. Maybe it's non-deterministic?