@Zoxc 2023-03-03
You could give this another perf run to see if it reproduces. I also added another optimization bringing instructions for allocating an `usize` from 8 down to 6.