@Zoxc 2023-03-04
This could use another perf run to see if more aggressive inlining helps recover the bootstrap improvement.