@Zoxc 2023-03-25
I retested it locally and it's still a performance improvement:
<table><tr><td rowspan="2">Benchmark</td><td colspan="1"><b>Before</b></th><td colspan="2"><b>After</b></th></tr><tr><td align="right">Time</td><td align="right">Time</td><td align="right">%</th></tr><tr><td>🟣 <b>clap</b>:check</td><td align="right">1.7190s</td><td align="right">1.7174s</td><td align="right"> -0.09%</td></tr><tr><td>🟣 <b>hyper</b>:check</td><td align="right">0.2621s</td><td align="right">0.2614s</td><td align="right"> -0.29%</td></tr><tr><td>🟣 <b>regex</b>:check</td><td align="right">0.9639s</td><td align="right">0.9552s</td><td align="right"> -0.90%</td></tr><tr><td>🟣 <b>syn</b>:check</td><td align="right">1.5532s</td><td align="right">1.5417s</td><td align="right"> -0.74%</td></tr><tr><td>🟣 <b>syntex_syntax</b>:check</td><td align="right">5.9043s</td><td align="right">5.8980s</td><td align="right"> -0.11%</td></tr><tr><td>Total</td><td align="right">10.4025s</td><td align="right">10.3737s</td><td align="right"> -0.28%</td></tr><tr><td>Summary</td><td align="right">1.0000s</td><td align="right">0.9957s</td><td align="right"> -0.43%</td></tr></table>

I think the good bootstrap result must have been a fluke as the last perf run has the same hot path.